+++ packages/devtools/client/composables/state-schema.ts
@@ -1,0 +1,6 @@
+export function useSchemaInput() {
+  return useSessionState<{
+    name?: string
+    input: string
+  } | null>('schema:input', null)
+}
